Add fallback to the optimized Sturm sequence builder
Fixes #1

This code has been added to address an issue with zeros while computing the remainder in the Sturm sequence builder. This section has not been optimized, profiled, or tested thoroughly. The fallback performs long polynomial division in the same way as if you were using a tabular method by hand. Only the last row of division has to be saved at each step.
To be verified: a break in the Sturm sequence should be treated as the sum of two individual sequences or as one shorter sequence?